Change callback to return an integer.

+++ library/src/net/simonvt/menudrawer/MenuDrawer.java
@@ -39,10 +39,11 @@ public abstract class MenuDrawer extends ViewGroup {
 
         /**
          * Called when the drawer slides.
-         * @param openRatio ratio for how open the menu is.
-         * @param offsetPixels current left position of the menu.
+         *
+         * @param openRatio    Ratio for how open the menu is.
+         * @param offsetPixels Current offset of the menu in pixels.
          */
-        void onDrawerSlide(float openRatio, float offsetPixels);
+        void onDrawerSlide(float openRatio, int offsetPixels);
     }
 
     /**
@@ -1176,7 +1177,7 @@ public abstract class MenuDrawer extends ViewGroup {
         return super.fitSystemWindows(insets);
     }
 
-    protected void dispatchOnDrawerSlide(float openRatio, float offsetPixels) {
+    protected void dispatchOnDrawerSlide(float openRatio, int offsetPixels) {
         if (mOnDrawerStateChangeListener != null) {
             mOnDrawerStateChangeListener.onDrawerSlide(openRatio, offsetPixels);
         }
