+++ samples/src/net/simonvt/menudrawer/samples/ListActivitySample.java
@@ -75,7 +75,7 @@ public class ListActivitySample extends ListActivity {
                 }
 
                 @Override
-                public void onDrawerSlide(float openRatio, float offsetPixels) {
+                public void onDrawerSlide(float openRatio, int offsetPixels) {
                     // Do nothing
                 }
             });
